+++ css/style.css
@@ -1,0 +1,33 @@
+
+a.nav-link{
+    color: white;
+    text-decoration: none;
+    font-family: Arial, Helvetica, sans-serif;
+    font-size: 20px;
+    padding: 5px;
+}
+div.nav-link-wrapper{
+    margin: 20px;
+    background-color: #0A66C2;
+    max-width: max-content;
+    border-radius: 2px;
+    padding: 5px;
+    transition: 0.2s;
+}
+
+div.nav-link-wrapper:hover{
+    background-color: #084480;
+}
+
+a.jobs{
+    text-decoration: none;
+    color: inherit;
+    font-family:Arial, Helvetica, sans-serif;
+    font-size: 30px;
+    font-weight: bold;
+}
+p.jobs{
+    max-width: 800px;
+    font-family:Arial, Helvetica, sans-serif;
+    font-size: 20px;
+}
