--- head.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html lang="sl">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Moje Delo</title>
-	<link rel="stylesheet" href="style.css">
-</head>
-<body>
