+++ html/head.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html lang="sl">
+<head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<title>Moje Delo</title>
+	<link rel="stylesheet" href="../css/style.css">
+</head>
+<body>
