+++ index.html
@@ -4,15 +4,15 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Start Page</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="css/style.css">
 </head>
 <body>
     <div class="nav-link-wrapper">
-        <a class="nav-link" href="moje_delo.html">Moje Delo</a>
+        <a class="nav-link" href="html/moje_delo.html">Moje Delo</a>
     </div>
 
     <div class="nav-link-wrapper">
-        <a class="nav-link" href="deloglsnik.html">Deloglasnik</a>
+        <a class="nav-link" href="html/deloglsnik.html">Deloglasnik</a>
     </div>
 </body>
 </html>
