+++ scripts/__init__.py
@@ -1,0 +1,2 @@
+from .deloglasnik import deloglasnik
+from .moje_delo import moje_delo
