+++ test.py
@@ -1,0 +1,3 @@
+import os
+
+print(os.listdir("html"))
